copilot-rest-python/copilot/requirements.txt
asgiref==3.8.1
Django==4.2.0
djangorestframework==3.15.1
sqlparse==0.5.0
